+++ docs/changelog.rst
@@ -13,6 +13,8 @@ New flavors:
 
 New fields for existing flavors:
 
+- Added MXPhoneNumberField model and form fields.
+
 - Added MXCLABEField model and form fields.
   (`gh-227 <https://github.com/django/django-localflavor/pull/227>`_).
 - Added AUTaxFileNumberField model and form fields.